decompose large files
