+++ integration_test/helpers/pod_auth_automator.dart
@@ -10,12 +10,12 @@
 library;
 
 import 'dart:convert';
-import 'dart:math';
 
-import 'package:crypto/crypto.dart';
 import 'package:puppeteer/puppeteer.dart';
 
 import '../test_constants.dart';
+import 'browser_automation_helpers.dart';
+import 'oauth_helpers.dart';
 
 /// Result of POD authentication automation.
 class AuthResult {
@@ -119,7 +119,7 @@ class PodAuthAutomator {
 
       // Perform dynamic client registration to get client_id.
       print('Registering OAuth client...');
-      final clientId = await _registerOAuthClient(page);
+      final clientId = await registerOAuthClient(page);
       if (clientId == null) {
         return AuthResult(
           success: false,
@@ -129,12 +129,12 @@ class PodAuthAutomator {
       print('✓ Client registered: $clientId');
 
       // Generate PKCE parameters.
-      final codeVerifier = _generateCodeVerifier();
-      final codeChallenge = _generateCodeChallenge(codeVerifier);
+      final codeVerifier = generateCodeVerifier();
+      final codeChallenge = generateCodeChallenge(codeVerifier);
       print('✓ Generated PKCE challenge');
 
       // Construct OAuth authorization URL with PKCE.
-      final authUrl = _buildAuthorizationUrl(clientId, codeChallenge);
+      final authUrl = buildAuthorizationUrl(clientId, codeChallenge);
       print('Navigating to OAuth authorization endpoint...');
       print('URL: ${authUrl.substring(0, 100)}...');
 
@@ -198,7 +198,7 @@ class PodAuthAutomator {
       print('Checking for consent screen...');
       if (page.url!.contains('/consent')) {
         print('Consent screen detected!');
-        final hasConsent = await _handleConsentScreen(page);
+        final hasConsent = await handleConsentScreen(page);
         if (hasConsent) {
           print('Consent granted, waiting for navigation...');
           try {
@@ -213,7 +213,7 @@ class PodAuthAutomator {
 
       // Handle security key input if present.
       print('Checking for security key prompt...');
-      final hasSecurityKey = await _handleSecurityKey(page, authSecurityKey);
+      final hasSecurityKey = await handleSecurityKey(page, authSecurityKey);
       if (hasSecurityKey) {
         print('Security key entered, waiting for callback...');
       }
@@ -255,7 +255,7 @@ class PodAuthAutomator {
 
       // Exchange authorization code for OAuth tokens.
       print('Exchanging authorization code for tokens...');
-      final tokenResponse = await _exchangeCodeForTokens(
+      final tokenResponse = await exchangeCodeForTokens(
         page,
         authorizationCode,
         clientId,
@@ -278,7 +278,7 @@ class PodAuthAutomator {
       final idToken = oauthTokens['id_token'] as String?;
       String? webId;
       if (idToken != null) {
-        webId = _extractWebIdFromIdToken(idToken);
+        webId = extractWebIdFromIdToken(idToken);
         if (webId != null) {
           print('✓ Extracted WebID: $webId');
         }
@@ -323,335 +323,9 @@ class PodAuthAutomator {
     }
   }
 
-  /// Handles OAuth consent screen if present.
-  static Future<bool> _handleConsentScreen(Page page) async {
-    try {
-      // Wait a moment for the consent screen to render.
-      await Future.delayed(const Duration(seconds: 2));
-
-      print('Looking for consent buttons...');
-
-      // Try to find and click the "Yes" button.
-      try {
-        // Look for button with text "Yes" (case-insensitive).
-        final buttons = await page.$$('button');
-        print('Found ${buttons.length} buttons on page');
-
-        for (final button in buttons) {
-          final text =
-              await page.evaluate('el => el.textContent', args: [button]);
-          final textStr = text.toString().trim().toLowerCase();
-
-          print('Button text: "$textStr"');
-
-          if (textStr == 'yes' ||
-              textStr == 'allow' ||
-              textStr == 'authorize' ||
-              textStr == 'consent') {
-            print('✓ Found consent button with text: "$text", clicking...');
-            await button.click();
-            return true;
-          }
-        }
-      } catch (e) {
-        print('Error finding consent button: $e');
-      }
-
-      // Alternative: try looking for input type=submit with "Yes" value.
-      try {
-        final submitInputs = await page.$$('input[type="submit"]');
-        for (final input in submitInputs) {
-          final value = await page.evaluate('el => el.value', args: [input]);
-          final valueStr = value.toString().trim().toLowerCase();
-
-          if (valueStr == 'yes' ||
-              valueStr == 'allow' ||
-              valueStr == 'authorize') {
-            print('✓ Found consent input with value: "$value", clicking...');
-            await input.click();
-            return true;
-          }
-        }
-      } catch (e) {
-        print('Error checking submit inputs: $e');
-      }
-
-      // Last resort: try the first submit button.
-      try {
-        final submitButtons = await page.$$('button[type="submit"]');
-        if (submitButtons.isNotEmpty) {
-          print('Trying first submit button as fallback...');
-          await submitButtons.first.click();
-          return true;
-        }
-      } catch (e) {
-        print('Error clicking submit button: $e');
-      }
-
-      print('No consent button found');
-      return false;
-    } catch (e) {
-      print('Error handling consent screen: $e');
-      return false;
-    }
-  }
-
-  /// Handles security key input if present.
-  static Future<bool> _handleSecurityKey(Page page, String securityKey) async {
-    try {
-      // Look for security key input.
-      final securityKeySelectors = [
-        'input[type="text"][placeholder*="key"]',
-        'input[type="text"][placeholder*="security"]',
-        'input[name="securityKey"]',
-        'input[id="securityKey"]',
-      ];
-
-      for (final selector in securityKeySelectors) {
-        try {
-          await page.waitForSelector(selector, timeout: Duration(seconds: 2));
-          await page.type(selector, securityKey);
-
-          // Click submit button.
-          await page.click('button[type="submit"]');
-          return true;
-        } catch (_) {
-          // Try next selector.
-        }
-      }
-
-      return false;
-    } catch (e) {
-      print('No security key prompt found or error handling it: $e');
-      return false;
-    }
-  }
-
   /// Pretty print tokens for debugging.
   static String formatTokens(Map<String, dynamic> tokens) {
     const encoder = JsonEncoder.withIndent('  ');
     return encoder.convert(tokens);
-  }
-
-  /// Exchanges authorization code for OAuth tokens.
-  ///
-  /// Returns a map with success status and tokens/error.
-  static Future<Map<String, dynamic>> _exchangeCodeForTokens(
-    Page page,
-    String authorizationCode,
-    String clientId,
-    String codeVerifier,
-  ) async {
-    try {
-      const tokenEndpoint = 'https://pods.dev.solidcommunity.au/.oidc/token';
-      const redirectUri = 'http://localhost:44007/';
-
-      // Prepare token request body.
-      final tokenRequest = {
-        'grant_type': 'authorization_code',
-        'code': authorizationCode,
-        'client_id': clientId,
-        'code_verifier': codeVerifier,
-        'redirect_uri': redirectUri,
-      };
-
-      // Use fetch API to exchange code for tokens.
-      final result = await page.evaluate(
-        '''
-        async (endpoint, data) => {
-          try {
-            // Convert data to URL-encoded form
-            const formBody = Object.keys(data)
-              .map(key => encodeURIComponent(key) + '=' + encodeURIComponent(data[key]))
-              .join('&');
-
-            const response = await fetch(endpoint, {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/x-www-form-urlencoded',
-              },
-              body: formBody,
-            });
-
-            if (!response.ok) {
-              const errorText = await response.text();
-              return {
-                success: false,
-                error: 'HTTP ' + response.status + ': ' + errorText
-              };
-            }
-
-            const json = await response.json();
-            return { success: true, tokens: json };
-          } catch (err) {
-            return { success: false, error: err.toString() };
-          }
-        }
-      ''',
-        args: [tokenEndpoint, tokenRequest],
-      );
-
-      if (result is Map && result['success'] == true) {
-        return {
-          'success': true,
-          'tokens': result['tokens'] as Map,
-        };
-      } else {
-        return {
-          'success': false,
-          'error': result['error']?.toString() ?? 'Unknown error',
-        };
-      }
-    } catch (e) {
-      print('Error exchanging code for tokens: $e');
-      return {
-        'success': false,
-        'error': e.toString(),
-      };
-    }
-  }
-
-  /// Extracts WebID from ID token JWT.
-  ///
-  /// ID tokens are JWTs with 3 parts: header.payload.signature
-  /// The payload contains the webid claim (or sub claim as fallback).
-  static String? _extractWebIdFromIdToken(String idToken) {
-    try {
-      // Split JWT into parts.
-      final parts = idToken.split('.');
-      if (parts.length != 3) {
-        print('Invalid ID token format');
-        return null;
-      }
-
-      // Decode payload (base64url).
-      final payload = parts[1];
-
-      // Add padding if needed for base64 decoding.
-      var normalized = payload.replaceAll('-', '+').replaceAll('_', '/');
-      while (normalized.length % 4 != 0) {
-        normalized += '=';
-      }
-
-      // Decode base64.
-      final decoded = utf8.decode(base64.decode(normalized));
-      final json = jsonDecode(decoded) as Map<String, dynamic>;
-
-      print('ID token payload: ${json.keys.toList()}');
-
-      // Extract webid claim (or sub as fallback).
-      final webId = json['webid'] as String? ?? json['sub'] as String?;
-      return webId;
-    } catch (e) {
-      print('Error extracting WebID from ID token: $e');
-      return null;
-    }
-  }
-
-  /// Registers an OAuth client dynamically with the Solid POD server.
-  ///
-  /// Returns the client_id if successful, null otherwise.
-  static Future<String?> _registerOAuthClient(Page page) async {
-    try {
-      // Navigate to the registration endpoint.
-      const registrationEndpoint =
-          'https://pods.dev.solidcommunity.au/.oidc/reg';
-
-      // Prepare registration request.
-      final registrationData = {
-        'client_name': 'MovieStar E2E Test Client',
-        'redirect_uris': ['http://localhost:44007/'],
-        'response_types': ['code'], // Authorization code flow only
-        'grant_types': ['authorization_code'],
-        'scope': 'openid profile',
-        'application_type': 'web',
-        'token_endpoint_auth_method':
-            'none', // Public client (no client secret)
-      };
-
-      // Use fetch API to register client.
-      final result = await page.evaluate(
-        '''
-        async (endpoint, data) => {
-          try {
-            const response = await fetch(endpoint, {
-              method: 'POST',
-              headers: {
-                'Content-Type': 'application/json',
-              },
-              body: JSON.stringify(data),
-            });
-
-            if (!response.ok) {
-              return { success: false, error: await response.text() };
-            }
-
-            const json = await response.json();
-            return { success: true, data: json };
-          } catch (err) {
-            return { success: false, error: err.toString() };
-          }
-        }
-      ''',
-        args: [registrationEndpoint, registrationData],
-      );
-
-      if (result is Map && result['success'] == true) {
-        final data = result['data'] as Map;
-        return data['client_id']?.toString();
-      } else {
-        print('Client registration failed: ${result['error']}');
-        return null;
-      }
-    } catch (e) {
-      print('Error registering OAuth client: $e');
-      return null;
-    }
-  }
-
-  /// Builds the OAuth authorization URL with proper parameters including PKCE.
-  static String _buildAuthorizationUrl(String clientId, String codeChallenge) {
-    const authEndpoint = 'https://pods.dev.solidcommunity.au/.oidc/auth';
-    const redirectUri = 'http://localhost:44007/';
-    const responseType = 'code'; // Authorization code flow
-    const scope = 'openid profile';
-
-    // Generate a random state for CSRF protection.
-    final state = DateTime.now().millisecondsSinceEpoch.toString();
-
-    // Build the URL with proper encoding including PKCE parameters.
-    final params = {
-      'response_type': responseType,
-      'client_id': clientId,
-      'redirect_uri': redirectUri,
-      'scope': scope,
-      'state': state,
-      'code_challenge': codeChallenge,
-      'code_challenge_method': 'S256', // SHA-256
-      'prompt': 'consent', // Force consent screen
-    };
-
-    final queryString = params.entries
-        .map(
-          (e) =>
-              '${Uri.encodeComponent(e.key)}=${Uri.encodeComponent(e.value)}',
-        )
-        .join('&');
-
-    return '$authEndpoint?$queryString';
-  }
-
-  /// Generates a random code verifier for PKCE.
-  static String _generateCodeVerifier() {
-    final random = Random.secure();
-    final values = List<int>.generate(32, (i) => random.nextInt(256));
-    return base64UrlEncode(values).replaceAll('=', '');
-  }
-
-  /// Generates a code challenge from the code verifier using SHA-256.
-  static String _generateCodeChallenge(String verifier) {
-    final bytes = utf8.encode(verifier);
-    final digest = sha256.convert(bytes);
-    return base64UrlEncode(digest.bytes).replaceAll('=', '');
   }
 }
